getTopFive() is Now Dynamic and Fixed Error
getTopFive() was not actually working upon running a test. It was
calculating the users with the highest averages, not the movies. I've
also tweaked the function to allow it to return highest averages in
ascending order up until the total amount of movies*.

*BUG: Cannot print element in last index if trying to sort entire list
from best to worst. Not sure why yet.